requirements.txt
python-telegram-bot==20.7
google-generativeai>=0.3.0
openai>=1.0.0
requests>=2.25.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
peewee>=3.17.0